moving types to specific folder
