+++ src/types/globals.d.ts
@@ -1,0 +1,7 @@
+// Global type declarations
+
+declare global {
+  interface Window {
+    customProperty?: string;
+  }
+}
