+++ src/types/hooks.d.ts
@@ -1,0 +1,6 @@
+// Types for hooks
+// Placeholder for hook-specific types if needed
+
+export interface EffectCallback {
+  (): void;
+}
