+++ src/types/services.d.ts
@@ -1,0 +1,6 @@
+// Types for services
+
+export interface ChatConfig {
+  botName: string;
+  botAvatar: string;
+}
